This is a lot of big changes that attempts to fix the problem that happens when createresultsfile is run after the first algorithm and the first algorithm wasn't able to run. It allows createresultsfile to be run on a subsequent algorithm. It does this while simultaneously avoiding the alternative problem that could occur if only a single algorithm were selected and it fails, if loadresultsfile isn't run in that circumstance, then the current results file gets overwritten with blank algorithms. To fix this problem I created loadresults to replace loadresultsfile.
Also updated the abd algorithm so that no results are stored if there is no HR signal, spo2 signal, or resp signal.

Also fixes a glitch where if a QRS algorithm is run first, then the result name stores the qrs display name, which shouldn't happen. This needs to be pushed to other branches.